+++ geotrek/trekking/serializers.py
@@ -131,6 +131,12 @@ class StructureSerializer(rest_serializers.ModelSerializer):
         fields = ('id', 'name')
 
 
+class ChildSerializer(TranslatedModelSerializer):
+    class Meta:
+        model = trekking_models.Trek
+        fields = ('id', )
+
+
 class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
                      AltimetrySerializerMixin, ZoningSerializerMixin,
                      TranslatedModelSerializer):
@@ -146,6 +152,7 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
     relationships = TrekRelationshipSerializer(many=True, source='published_relationships')
     treks = CloseTrekSerializer(many=True, source='published_treks')
     source = RecordSourceSerializer()
+    children = rest_serializers.Field(source='published_children_id')
 
     # Idea: use rest-framework-gis
     parking_location = rest_serializers.SerializerMethodField('get_parking_location')
@@ -183,7 +190,8 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
                   'web_links', 'is_park_centered', 'disabled_infrastructure',
                   'parking_location', 'relationships', 'points_reference',
                   'poi_layer', 'information_desk_layer', 'gpx', 'kml', 'source',
-                  'type1', 'type2', 'category', 'structure', 'treks') + \
+                  'type1', 'type2', 'category', 'structure', 'treks',
+                  'parent', 'children') + \
             AltimetrySerializerMixin.Meta.fields + \
             ZoningSerializerMixin.Meta.fields + \
             PublishableSerializerMixin.Meta.fields + \