+++ docs/changelog.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
 **New features**
 
+* Itinerancy (parent/children treks)
 * Allow to choose ordering of categories for Geotrek-Rando
 * Bootstrap grid editor for flatpages